+++ changelog.md
@@ -1,1 +1,1 @@
-- Updated Forge version
+- Fixed experience bar speed and fuel indicators